Skip :verify_authenticity_token filter on OpenID callback
When a user returns from the OpenID server and hits the sso callback,
the session is reset, and so the user_return_to URL is lost.  As a
result, the callback redirects the user to the default location,
root_url, instead of the original URL they tried to access when they
began authentication.

Skip the check of the authenticity_token: it's not useful in this case
because the OpenID server won't be sending a token.