Add paid_content which requires permission granted by O'Reilly to access it
We want to authorize access to paid content hosted on a web site using
O'Reilly Media Manager, the system that currently tracks ownership of
intellectual property such as ebooks.

Demonstrate this with the members permissions by fetching the
appropriate permission and then parsing the RDF to ensure it is not
empty.