Use membersstage with HTTP Basic Auth in development and production
We now have a mechanism to allow access to the permissions API based
on HTTP Basic Auth instead of just by client IP address.

Configure the app for a password to be set in an environment variable
(which works with heroku config in production), and point the app at
membersstage, where the new access control mechanism is currently
staged.