Fetch product titles from OPMI
When listing the products that a user has permission to access, a title
is far more meaningful than a product URI.

Use the public OPMI API to access metadata about a product, and extract
the om:customerTitle for use as the product's title.  Note that, because
OPMI only shows public products, some products recognized in media
manager will find their title in OPMI.  So it's possible to have a nil
title returned.