Update Permission.can_access to use new permission service
The old permissions API for checking authorization to a particular
product was over-complicated because it exposed an RDF response when a
simple boolean response was sufficient.  We should now upgrade to the
new API.

Use the new API by:

* Adding config.oreilly.permissions_service_url
* Extract an HttpApi module from RdfApi that wraps Net::HTTP without RDF
* Change can_access and permitted_products to accept objects rather than
  strings.  can_access needs multiple fields from the Product, and
  authorize! probably shouldn't need to know the details of which user
  fields are passed to the API.

Leave permitted_products to use the old RDF API because there is no
replacement for it (yet).