Improve error message to include ISBN when not authorized
We're changing the product that confers access to the paid_content page,
so anyone who was setup with permissions before will need to get new
media permissions for the new title to see the authorized use case.

Make it easy to know which product to add by supplying right in the
error message.