Hide HTTP 404 errors within the Product#title accessor
Some products we get from the Permissions#permitted_products API are
not published in OPMI, so we cannot get their titles.  We currently
raise an unhandled exception in this case.

Handle the exception by treating it as a nil title: this kind of
accessor should probably not leak HTTP exceptions.